fixed minor spacing and moved footer
