+++ frontend/src/components/Footer.tsx
@@ -1,45 +1,16 @@
 "use client"
 
 import React, { useRef } from "react";
-import { TextGenerateEffect } from "@/components/ui/text-generate-effect";
-import Button from "@/components/ui/Button";
-import { motion } from "framer-motion";
-import { useInView } from "framer-motion";
 
-const FooterSection = () => {
-    const ref = useRef(null);
-    const isInView = useInView(ref, { once: false, amount: 0.3 });
+
+
+const Footer = () => {
+
 
 
     return (
         <>
-            {/* Shadow line with animation */}
-            <div ref={ref} className="relative w-full h-[4rem] flex justify-center overflow-hidden">
-                <motion.span
-                    className="absolute shadow-lg shadow-blue-500/50 border-t border-gradient-to-r from-blue-500 to-teal-400"
-                    style={{
-                        boxShadow: " 0 0 40px 10px rgba(56, 191, 248, 0.9),  0 0 40px 10px rgba(45, 212, 191, 0.9)",
-                    }}
-                    initial={{ width: "0%" }}
-                    animate={isInView ? { width: "100%" } : { width: "0%" }}
-                    transition={{ duration: 1.2, ease: "easeOut" }}
-                />
-            </div>
 
-            <div className="container mx-auto px-6 text-center mb-10" >
-                <h2 className="text-3xl font-bold text-gray-800 mb-6">
-                    <TextGenerateEffect words={'Start Your Language Learning Journey Today'} className="text-3xl bg-gradient-to-r from-blue-500 to-teal-400 bg-clip-text text-transparent" />
-                </h2>
-                <motion.div
-                    initial={{ y: "50%", opacity: "0%" }}
-                    animate={isInView ? { y: "0%", opacity: "100%" } : { y: "50%", opacity: "0%" }}
-                    transition={{ duration: 1, ease: "easeOut" }}
-
-                >
-                    <Button text="Sign Up Free" />
-                </motion.div>
-
-            </div>
             <footer className="bg-gray-800 text-white py-12">
                 <div className="container mx-auto px-6">
                     <div className="flex flex-col md:flex-row justify-between mb-8">
@@ -98,4 +69,4 @@ const FooterSection = () => {
     )
 }
 
-export default FooterSection;
+export default Footer;
